Move view helpers to helpers file.
